Update TargetFramework to netstandard2.1
- Use file scoped namespace declarations
- Update LangVersion to 13.0
- Change the source directory structure
- Remove CircleCI, AppVeyor, Azure Pipelines
- Add workflow files for GitHub Actions
- Update dependencies
- Update documentation
- Bump version to 2.0.0
