+++ .gitignore
@@ -1,7 +1,6 @@
 .vs/
 Coverlet-html/
 CuiMallet.Test/bin/
-CuiMallet.Test/coverage.opencover.xml
 CuiMallet.Test/dcx/
 CuiMallet.Test/obj/
 CuiMallet/bin/
