+++ CuiMallet/Cui/Impl/Resources.ja-JP.resx
