+++ CuiMallet/CuiMallet.csproj
@@ -1,10 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace>Maroontress.CuiMallet</RootNamespace>
+    <TargetFramework>netstandard2.1</TargetFramework>
+    <RootNamespace>Maroontress</RootNamespace>
     <AssemblyName>Maroontress.CuiMallet</AssemblyName>
-    <LangVersion>8.0</LangVersion>
+    <LangVersion>13.0</LangVersion>
     <Nullable>enable</Nullable>
     <DocumentationFile>dcx\CuiMallet.xml</DocumentationFile>
   </PropertyGroup>
@@ -21,7 +21,7 @@
     <Copyright>Copyright (c) 2019 Maroontress Fast Software</Copyright>
     <PackageTags>CLI, csharp, csharp8, dotnet-standard, command-line-parser, command-line-arguments</PackageTags>
     <NoPackageAnalysis>true</NoPackageAnalysis>
-    <Version>1.0.0.0</Version>
+    <Version>2.0.0</Version>
     <RepositoryType />
     <Company>Maroontress Fast Software</Company>
   </PropertyGroup>
@@ -43,9 +43,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="StyleChecker" Version="1.0.25" PrivateAssets="all" />
+    <PackageReference Include="StyleChecker" Version="2.0.1" PrivateAssets="all" />
     <PackageReference Include="StyleChecker.Annotations" Version="1.0.1" />
-    <PackageReference Include="System.Collections.Immutable" Version="1.7.1" />
+    <PackageReference Include="System.Collections.Immutable" Version="9.0.0" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.164" PrivateAssets="all" />
   </ItemGroup>
 
@@ -65,12 +65,12 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="Maroontress\Cui\Impl\Resources.Designer.cs">
+    <Compile Update="Cui\Impl\Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Update="Maroontress\Cui\Impl\Resources.resx">
+    <EmbeddedResource Update="Cui\Impl\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
       <CustomToolNamespace>Maroontress.Cui.Impl</CustomToolNamespace>
