--- CuiMallet/Maroontress/Cui/InvalidOptionSchemaException.cs
@@ -1,47 +1,0 @@
-namespace Maroontress.Cui
-{
-    using System;
-
-    /// <summary>
-    /// Represents errors on the option schema.
-    /// </summary>
-    public sealed class InvalidOptionSchemaException : Exception
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see
-        /// cref="InvalidOptionSchemaException"/> class.
-        /// </summary>
-        public InvalidOptionSchemaException()
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see
-        /// cref="InvalidOptionSchemaException"/> class.
-        /// </summary>
-        /// <param name="message">
-        /// The message that describes the error.
-        /// </param>
-        public InvalidOptionSchemaException(string message)
-            : base(message)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see
-        /// cref="InvalidOptionSchemaException"/> class.
-        /// </summary>
-        /// <param name="message">
-        /// The error message that explains the reason for the exception.
-        /// </param>
-        /// <param name="innerException">
-        /// The exception that is the cause of the current exception, or a null
-        /// reference if no inner exception is specified.
-        /// </param>
-        public InvalidOptionSchemaException(
-            string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
-    }
-}
