--- appveyor.yml
@@ -1,10 +1,0 @@
-version: '{build}'
-image: Visual Studio 2019
-configuration: Debug
-platform: Any CPU
-init:
-- cmd: git config --global core.autocrlf true
-before_build:
-- cmd: nuget restore
-build:
-  verbosity: minimal
