--- azure-pipelines.yml
@@ -1,39 +1,0 @@
-trigger:
-  branches:
-    include:
-    - '*'
-
-pool:
-  vmImage: 'windows-2019'
-
-steps:
-- checkout: self
-  lfs: true
-- task: UseDotNet@2
-  displayName: install .NET Core SDK
-  inputs:
-    packageType: sdk
-    version: 3.1.300
-- task: CmdLine@2
-  inputs:
-    script: dotnet tool install -g BomSweeper.GlobalTool
-    displayName: install BomSweeper.GlobalTool
-- task: CmdLine@2
-  inputs:
-    script: bomsweeper '**/*.cs' '**/*.csproj' '**/*.sln'
-    displayName: BOM check
-    failOnStderr: true
-- task: DotNetCoreCLI@2
-  displayName: dotnet build
-  inputs:
-    command: 'build'
-- task: DotNetCoreCLI@2
-  displayName: dotnet test
-  inputs:
-    command: 'test'
-    arguments: '-p:CollectCoverage=true /p:CoverletOutputFormat=cobertura --no-build CuiMallet.Test'
-- task: PublishCodeCoverageResults@1
-  displayName: publish coverage report
-  inputs:
-    codeCoverageTool: Cobertura
-    summaryFileLocation: 'CuiMallet.Test/coverage.cobertura.xml'
